% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Palmer Penguins dataset}
\format{
A data frame with 344 rows and 8 variables:
\describe{
\item{species}{Penguin species}
\item{island}{Island name}
\item{bill_length_mm}{Bill length}
\item{bill_depth_mm}{Bill depth}
\item{flipper_length_mm}{Flipper length}
\item{body_mass_g}{Body mass}
\item{sex}{Sex}
\item{year}{Year}
}
}
\source{
\url{https://allisonhorst.github.io/palmerpenguins/}
}
\usage{
penguins
}
\description{
A cleaned dataset of penguin measurements from the Palmer Station LTER.
}
\keyword{datasets}
